+++ workspaces/mi/mi-extension/src/rpc-managers/mi-visualizer/rpc-manager.ts
@@ -99,8 +99,7 @@ import { findMultiModuleProjectsInWorkspaceDir } from "../../util/migrationUtils
 import { MILanguageClient } from "../../lang-client/activator";
 import { reorderModulesByBuildOrder } from "../../debugger/pomResolver";
 import { buildDeployExtraArgs, executeRemoteDeployTask } from "../../debugger/debugHelper";
-import { getAnthropicClient, ANTHROPIC_HAIKU_4_5 } from "../../ai-features/connection";
-import { generateText } from "ai";
+import { MIAIPanelRpcManager } from "./../../rpc-managers/ai-features/rpc-manager";
 
 Mustache.escape = escapeXml;
 
@@ -1021,46 +1020,9 @@ export class MiVisualizerRpcManager implements MIVisualizerAPI {
             return originalUrl;
         }
     }
-    
+
     async getMcpToolSuggestion(params: McpToolSuggestionRequest): Promise<McpToolSuggestionResponse> {
-        const { toolName, operationMethod, operationPath, operationSummary } = params;
-
-        const contextLines: string[] = [`Tool name: ${toolName}`];
-        if (operationMethod) contextLines.push(`HTTP method: ${operationMethod}`);
-        if (operationPath) contextLines.push(`HTTP path: ${operationPath}`);
-        if (operationSummary) contextLines.push(`Operation summary: ${operationSummary}`);
-        const context = contextLines.join('\n');
-
-        const prompt = `You are helping to configure an MCP (Model Context Protocol) tool for a WSO2 Micro Integrator project.
-
-Given the following tool context:
-${context}
-
-Generate:
-1. A clear, one-sentence description of what this tool does (suitable as the MCP tool description shown to an AI assistant).
-2. A JSON input schema using shorthand notation {"paramName": "type"} where type is one of: string, number, boolean, integer. Always infer typical parameters based on the tool name and context — for example, get_weather would have {"city": "string", "units": "string"}, create_order would have {"customerId": "string", "amount": "number"}.
-
-Respond ONLY with a JSON object in this exact format, no other text:
-{"description": "...", "inputSchema": {"param1": "type1", "param2": "type2"}}`;
-
-        try {
-            const model = await getAnthropicClient(ANTHROPIC_HAIKU_4_5);
-            const { text } = await generateText({
-                model: model,
-                prompt: prompt,
-                temperature: 0.3,
-                maxOutputTokens: 500,
-            });
-
-            const cleaned = text.trim().replace(/^```json\s*/i, '').replace(/```\s*$/, '');
-            const parsed = JSON.parse(cleaned);
-            return {
-                description: parsed.description || '',
-                inputSchema: JSON.stringify(parsed.inputSchema || {}),
-            };
-        } catch (error: any) {
-            vscode.window.showErrorMessage(`Fill With AI failed: ${error?.message ?? 'Unknown error'}`);
-            return { description: '', inputSchema: '{}' };
-        }
+        const aiManager = new MIAIPanelRpcManager(this.projectUri);
+        return aiManager.getMcpToolSuggestion(params);
     }
 }
